Added integration tests for TrustsIntegrationTests.
Unconvered some difficult behaviour with the SmartData table,
EntityFramework, and NBuilder.

The SmartData table has quite a few restrictions on values for columns,
in particular the Decimal and Numeric columns. The values that NBuilder
generated, even when using min/max values that should have been correct,
were causing overflow errors when inserting into the database. This
happened intermittently, so made for quite flaky tests.

Manually specifying the values, and using static Decimal values for the
decimal fields seems to have solved this; the other fields are all
randomised so the generated objects still have enough entropy to be
valuable in test triangulation.

Moved the function to a Generators util class, rather than repeating it
all over the place.