Add UseCase to get Establishments by TrustUid
We want to add an endpoint to get Trusts by UKPRNs. As part of this we we will
need to fetch the related Establishments.

This adds a UseCase to get Establishments by TrustUids, so that we can get the
Establishments related to the Trust by performing as few database queries as
possible. The fetching of Establishments by multiple URNs or Trust UIDs is
squashed into a single `GetEstablishments` UseCase to match the existing
`GetEstablishment` UseCase