Add controller action to get multiple Trusts by UKPRNs
This adds a `bulk` endpoint in the format
`/v2/trusts/bulk?Ukprn=12345678&Urn=234566789"` so that multiple trusts can be
retrieved in a single request.

There is no standard for the correct way to pass multiple values to an
endpoint, so I have opted for the easiest built-in method (which also matches
the `bulk` Establishments action)

The endpoint will only return `not found` when no trusts are found for
all of the UKPRNs passed, otherwise the trusts that were found will be
returned.

This also tidies up any shared database clean up into a `Dispose` method in the
integration tests