Add UseCase to get Trusts by UKPRNs
Currently, in the Complete application, we fetch trust data for each project on
our project index page as a separate API request. This is at times very slow,
and while we have explored various options to speed this up in the client, we
think adding a `bulk` endpoint to the API may be the better choice.

This adds a use case for getting Trusts by a list of UKPRNs. 

So that we aren't performing a database query for each related piece of data
per UKPRN (`Group`, `Trust`, `Establishments` etc.), this loads lists of each
into memory before mapping them to the result.