Remove the Get Establishment by URN UseCase
We have recently added a Get Establishments By URNs UseCase. We can call this
when we want to retrieve a single Establishment in order to DRY up our
UseCases. 

We have seen some performance improvements on the new UseCase, so this will
bring those to the single resource endpoints, as well as making the code more
maintainable.

This includes some additional changes:
- The get Establishments by URNs endpoint and get Trusts by UKPRNs endpoint 
  will now return Census data.
- The Census Gateway is updated, so that we do not need to read the CSV for each
  URN when we need census data.
- Removes the separate `GetEstablishment` tests for each piece of related data and
  squashes them into one, as they aren't providing any value by being separate.