Remove GetEstablishmentByUkprn UseCase
We have recently added a Get Establishments By UKPRNs UseCase. We can call this
when we want to retrieve a single Establishment in order to DRY up our
UseCases.

We have seen some performance improvements on the new UseCase, so this will
bring those to the single resource endpoints, as well as making the code more
maintainable.

The Establishment Gateway is updated, so that we do not need to perform a query
for each separate UKPRN.